Review #8 §13/§19: reject FlowServoAsValve; structural soft-limit/level checks
§13: FlowServoAsValve is declared but not wired — MainApp drives a separate,
unattached gate servo while the flow servo runs independently, so the valve
closes nothing. Reject it as UNSUPPORTED_BACKEND and repoint its preset at a
real ServoValve gate so presets stay validation-clean.

§19: add HardwareResourceValidator::claimBound (BOUND_INVALID) and validate soft
limits within [0, travelMm] and ordered, plus source min01 <= max01. Use a
64-bit product for the homing phaseTimeoutMs ms->us scale so a large timeout
cannot overflow.

+++ esp32/esp32_slide_whistle/core/SlideActuators.h
@@ -245,8 +245,10 @@ protected:
 
     void homingStep(float dt, uint32_t nowUs) override {
         if (!havePhaseTime_) { phaseStartUs_ = nowUs; havePhaseTime_ = true; }
-        // per-phase timeout — never spins forever (correction #4/#5)
-        if (elapsed_u32(nowUs, phaseStartUs_) > cfg_.stepper.phaseTimeoutMs * 1000u) {
+        // per-phase timeout — never spins forever (correction #4/#5). 64-bit
+        // product so a large phaseTimeoutMs cannot overflow the ms→us scale
+        // (review #8 §19).
+        if (elapsed_u32(nowUs, phaseStartUs_) > (uint64_t)cfg_.stepper.phaseTimeoutMs * 1000u) {
             fault_ = FaultCode::HomingTimeout;
             state_ = MotionState::Fault;
             if (sink_) sink_->enableDriver(false);
